// packages/codex-app-server/src/_generated/InitializeParams.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClientInfo } from "./ClientInfo.ts";
import type { InitializeCapabilities } from "./InitializeCapabilities.ts";

export type InitializeParams = { clientInfo: ClientInfo, capabilities: InitializeCapabilities | null, };